fix(ayokoding-www): enable static content rendering
