+++ specs/apps/ayokoding/behavior/ayokoding-www/gherkin/content/static-delivery.feature
@@ -1,0 +1,33 @@
+Feature: Statically delivered content pages
+
+  As a reader visiting AyoKoding
+  I want content pages to be delivered from the CDN with the correct document language
+  So that courses load quickly in my selected locale
+
+  Background:
+    Given the app is running
+
+  @e2e
+  Scenario: A content page is prerendered at build time
+    Given the ayokoding-www site is built and deployed
+    When the build output manifest is inspected
+    Then the prerendered route count is at least two thousand
+    And the content catch-all route is not marked as dynamically rendered
+
+  @e2e
+  Scenario: A repeat request to a content page is served from the CDN
+    Given a visitor has already requested a course lesson URL
+    When the same URL is requested again
+    Then the response is served from the CDN cache
+    And the response does not carry a no-store cache directive
+
+  @e2e
+  Scenario Outline: The document language reflects the content-page locale
+    Given a visitor opens a content page in the "<locale>" locale
+    When the content page renders
+    Then the html element declares the "<language_code>" language code
+
+    Examples:
+      | locale | language_code |
+      | en     | en            |
+      | id     | id            |
